Use colours.inc. Add more information and a test for an empty set of files.

+++ scripts/remove-all-deleted-ebuilds.sh
@@ -1,8 +1,16 @@
 #!/bin/sh
 
+source ./scripts/colours.inc
+
+printf ">> ${YELLOW}REMOVING ALL DELETED EBUILDS${NC}\n"
+
 EBUILDS=$(git status |grep ebuild |grep deleted|awk '{ print $2 }')
 
-echo ${EBUILDS}
+if [ -z ${EBUILDS} ]; then
+    printf ">> ${RED}None Found${NC}.\n"
+else
+    echo ${EBUILDS}
 
-git rm ${EBUILDS}
-git commit -m "Remove all old ebuilds." -v ${EBUILDS}
+    git rm ${EBUILDS}
+    git commit -m "Remove all old ebuilds." -v ${EBUILDS}
+fi
